git-svn-id: https://svn.r-project.org/R/branches/R-defs@78232 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/engine.c
@@ -1815,9 +1815,9 @@ void GEText(double x, double y, const char * const str, cetype_t enc,
 					}
 					done = TRUE;
 				    } else if (enc2 == CE_UTF8) {
-					size_t used;
+					int used;
 					wchar_t wc;
-					while ((used = utf8toucs(&wc, ss)) > 0) {
+					while ((used = (int) utf8toucs(&wc, ss)) > 0) {
 					    if (IS_HIGH_SURROGATE(wc))
 					    	GEMetricInfo(-(int)utf8toucs32(wc, ss), gc, &h, &d, &w, dd);
 					    else
@@ -2634,9 +2634,9 @@ void GEStrMetric(const char *str, cetype_t enc, const pGEcontext gc,
                         s += used; n -=used;
                     }
                 } else if (enc2 == CE_UTF8) {
-                    size_t used;
+                    int used;
                     wchar_t wc;
-                    while ((used = utf8toucs(&wc, s)) > 0) {
+                    while ((used = (int)utf8toucs(&wc, s)) > 0) {
                     	if (IS_HIGH_SURROGATE(wc))
                     	    GEMetricInfo(-utf8toucs32(wc, s), gc, &asc, &dsc, &wid, dd);
                     	else
@@ -2695,9 +2695,9 @@ void GEStrMetric(const char *str, cetype_t enc, const pGEcontext gc,
                         s += used; n -=used;
                     }
                 } else if (enc2 == CE_UTF8) {
-                    size_t used;
+                    int used;
                     wchar_t wc;
-                    while ((used = utf8toucs(&wc, s)) > 0) {
+                    while ((used = (int)utf8toucs(&wc, s)) > 0) {
                         if (IS_HIGH_SURROGATE(wc))
                             GEMetricInfo(-utf8toucs32(wc, s), gc, &asc, &dsc, &wid, dd);
                         else
